Templajt klasa za vracanje odgovora na front.

+++ XWS_2018_Backend/src/main/java/com/ftn/WebXML2018/XWS_2018_Backend/repository/CountryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ftn.WebXML2018.XWS_2018_Backend.entity.Country;
 
-public interface CountryRepository extends JpaRepository<Country, Integer>{
+public interface CountryRepository extends JpaRepository<Country, Long>{
 
 }
