+++ XWS_2018_Backend/src/main/java/com/ftn/WebXML2018/XWS_2018_Backend/service/RegisteredUserService.java
@@ -6,6 +6,6 @@ public interface RegisteredUserService {
 	
 	public RegisteredUser saveRegUser(RegisteredUser user);
 	
-	public RegisteredUser getRegUser(Integer id);
+	public RegisteredUser getRegUser(Long id);
 
 }
