+++ XWS_2018_Backend/src/main/java/com/ftn/WebXML2018/XWS_2018_Backend/service/UserService.java
@@ -6,6 +6,6 @@ public interface UserService {
 	
 	public User saveUser(User user);
 	
-	public User getUser(Integer id);
+	public User getUser(Long id);
 
 }
