+++ XWS_2018_Backend/src/main/java/com/ftn/WebXML2018/XWS_2018_Backend/serviceImpl/CountryServiceImpl.java
@@ -16,7 +16,7 @@ public class CountryServiceImpl implements CountryService{
 	private CountryRepository countryRepository;
 
 	@Override
-	public Country getOne(Integer id) {
+	public Country getOne(Long id) {
 		
 		return countryRepository.findOne(id);
 	}
