+++ XWS_2018_Backend/src/main/java/com/ftn/WebXML2018/XWS_2018_Backend/serviceImpl/RegisteredUserServiceImpl.java
@@ -20,7 +20,7 @@ public class RegisteredUserServiceImpl implements RegisteredUserService{
 	}
 
 	@Override
-	public RegisteredUser getRegUser(Integer id) {
+	public RegisteredUser getRegUser(Long id) {
 		
 		return registeredUserRepository.findOne(id);
 	}
